added survey_template convert

+++ DynamicSurvey.Server.DAL/Entities/SurveyField.cs
@@ -8,14 +8,14 @@ namespace DynamicSurvey.Server.DAL.Entities
 {
 	public class SurveyField
 	{
-		public int Id { get; set; }
+		public decimal Id { get; set; }
 		public string Label { get; set; }
 
 		public string FieldType { get; set; }
 
 		public string[] DefaultValues { get; set; }
 
-		public int? GroupId { get; set; }
+		public decimal? GroupId { get; set; }
 
 		public string UserAnswer { get; set; }
 	}
